Error on schema with attribute names duplicated in types (#152)
This was the original behavior, but it regressed when schema type
parsing was rewritten in #132. The regression was not noticed due to a
faulty test case. This commit also fixes the test case, and adds new
test cases for other places where a duplicate key might go unnoticed.

(cherry picked from commit 587412c5f260f7888115f45867f750a752f1c281)