Fix record LUB that was Open when it could be Closed
The record LUB was incorrectly made Open even though the LUB attribtues
keys match the input record attributes exactly. This happend only when
the records were not in a subtype releation becuase at least one pair
of corresponding attirbutes in the record were not in a subtype
relation, but the pair of attributes did have a least upper bound.
For example, `{a: true}` and `{a: false}`.