Fix record LUB that was Open when it could be Closed (#165)
The record LUB was incorrectly made Open even though the LUB attributes keys
match the input record attributes exactly. This happens only when the records
were not in a subtype relation because at least one pair of corresponding
attributes in the record were not in a subtype relation, but the pair of
attributes did have a least upper bound. For example, {a: true} and {a: false}.