Fix panic is cedar_policy::PolicySet::link (#203)
Removes an `expect` in this function which was known to panic when the
function was called with a `template_id` corresponding to a static
policy. An error result in returned instead of panicking.

Change return type of PolicySet::link to avoid an expect.
This `expect` did not panic, but this change eliminates the possibility.