Typecheck template based on possible instantiation types (#274)
Rather than typechecking a template once using AnyEntity for the type of the
slot, we can typecheck it once for each possible entity type that could
instantiate that slot.

This does not affect current permissive validation. It is instead intended to
support strict typechecking templates after the coming strict validation
changes. It will also support more precise template typechecking if template
slots are allowed in policy bodies in the future.