Fix issue #112
Fixes the evaluation order of operands to `>` and `>=` so that an error
in the left operand will always be reported instead of any error which
may exist in the right operand.

`a > b` is defined as `!(a <= b)` instead of `b < a` and `a >= b` is
defined as `!(a > b)` instead of `b <= a`.