Fix `PolicySet::link` to avoid mutating policy set on link fail
If linking failed because the template id corresponded to a static
policy, `PolicySet::link` would correctly return an error, but the
error was returned after calling `ast::PolicySet::link`. When given a
static policy id, this function updated the links HashMap to add an
entry for the target policy id which was not removed when returning an
error.