Fix typechecking for `in Action::"..."`
The partial schema PR inadvertantly changed the behavior of, e.g.,
`princpal in Action::"view"` to be an error, claiming that the `Action`
entity type is not defined. It should instead recognize that actions may
only be `in` other actions and give the expression type `False`.

Our unit tests check the `User::"alice" in Action::"view"` case, but did
not test the case when the left side of the `in` was a var.