Avoid claiming to expect tokens that aren't valid in any AST
There are tokens we include in our grammar to support better error
reporting in the cst to ast transform that will _always_ cause the
transformation to fail. E.g, `%` is not a valid Cedar operator. Filter
these tokens out of error messages listing what tokens are expected
after a parse error.

Signed-off-by: John Kastner <jkastner@amazon.com>