Add source location to cedar-wasm validate result
Adds the source location to the cedar-wasm output struct.
Additionally removes the use of `#[serde(untagged)]` in the validate
result, in favor of a "type" tag field.
The use of `untagged` made the resulting TypeScript result object difficult
to use when working with the generated TypeScript SDK.
Type tags are (IMO) fairly standard in the TypeScript ecosystem and are
familiar to TS developers, despite the perhaps marginal serialization overhead.

Signed-off-by: chrnorm <17420369+chrnorm@users.noreply.github.com>