Make `parse_policy_or_template_to_est` for CST that aren't valid AST
The EST conversion does not apply the checks that we apply CST to EST
conversion, and we don't want to duplicate all of them. For example,
this function succeeded for expression using extension function with an
incorrect call style like `"1".ip()` (cedar-policy/cedar-spec#341).

Effects `policy_text_to_json` in the WASM bindings.

Signed-off-by: John Kastner <jkastner@amazon.com>