Update ast type annotation to never eliminate expressions
Even when this should be possible, like when there are singleton boolean
types in `if`, `&&`, and `||` expression, it can cause trouble for
downstream analysis. In particular this change fixes a bug in entity
manifests where a slice based on a manifest would not capture attributes
access only `if` expression guards when the guard had a singleton boolean
type.

Any downstream analysis that wants the reduced AST should instead
inspect the type of expression and handle any singleton boolean types as
appropriate for that analysis.

Signed-off-by: John Kastner <jkastner@amazon.com>