Don't allow text-direction characters in literals
Replace text-direction marker codepoints in test string
literals with unicode escapes. This removes the need for
a crate-level allow of the `text_direction_codepoint_in_literal`
lint.

In rust 1.89 this lint was changed to function only crate-wide,
to better handle string literals produced by macro expansion.
As such it is not possible to limit the scope of the `allow`
only to test strings.

A downside of using escape sequences are that they can
make mixed rtl language text unreadable. That is not an
issue here.

A more significant downside is that escape sequences conflict
with raw strings, so the test policy stanzas must also be
manually escaped for other meaningful punctuation like
quotation marks. This is tedious and error-prone, but
considered the better option against the malicious obfuscation
attacks possible where text-direction markers interaction
with ltr text.

Resolves #1777.

Signed-off-by: Ralph Giles <giles@thaumas.net>